+++ pyemma/msm/estimators/maximum_likelihood_msm.py
@@ -812,7 +812,7 @@ class _MSMEstimator(_Estimator, _MSM):
     ################################################################################
 
     def cktest(self, nsets, memberships=None, mlags=10, conf=0.95, err_est=False,
-               n_jobs=1, show_progress=True):
+               n_jobs=None, show_progress=True):
         """ Conducts a Chapman-Kolmogorow test.
 
         Parameters
@@ -833,7 +833,7 @@ class _MSMEstimator(_Estimator, _MSM):
             compute errors also for all estimations (computationally expensive)
             If False, only the prediction will get error bars, which is often
             sufficient to validate a model.
-        n_jobs : int, default=1
+        n_jobs : int, default=None
             how many jobs to use during calculation
         show_progress : bool, optional
             Show progress bars for calculation?