+++ pyemma/plots/networks.py
@@ -157,14 +157,20 @@ class NetworkPlot(object):
         # number of nodes
         n = len(self.pos)
         # get bounds and pad figure
-        xmin = _np.min(self.pos[:, 0])
-        xmax = _np.max(self.pos[:, 0])
+        if self.ax is None:
+            xmin = _np.min(self.pos[:, 0])
+            xmax = _np.max(self.pos[:, 0])
+        else:
+            xmin, xmax = self.ax.get_xlim()
         Dx = xmax - xmin
         xmin -= Dx * figpadding
         xmax += Dx * figpadding
         Dx *= 1 + figpadding
-        ymin = _np.min(self.pos[:, 1])
-        ymax = _np.max(self.pos[:, 1])
+        if self.ax is None:
+            ymin = _np.min(self.pos[:, 1])
+            ymax = _np.max(self.pos[:, 1])
+        else:
+            ymin, ymax = self.ax.get_ylim()
         Dy = ymax - ymin
         ymin -= Dy * figpadding
         ymax += Dy * figpadding
@@ -186,10 +192,9 @@ class NetworkPlot(object):
             figsize = (Dx / Dy * max_height, max_height)
         if self.ax is None:
             logger.debug("creating new figure")
-            fig = _plt.figure(None, figsize=figsize)
-            self.ax = fig.add_subplot(111)
+            fig, self.ax = _plt.subplots(figsize=figsize)
         else:
-            fig = self.ax.figure
+            fig = self.ax.get_figure()
             window_extend = self.ax.get_window_extent()
             axes_ratio = window_extend.height / window_extend.width
             data_ratio = (ymax - ymin) / (xmax - xmin)
@@ -579,7 +584,7 @@ def plot_network(
     weights, pos=None, xpos=None, ypos=None, state_sizes=None, state_scale=1.0,
     state_colors='#ff5500', state_labels='auto', arrow_scale=1.0, arrow_curvature=1.0,
     arrow_labels='weights', arrow_label_format='%2.e', max_width=12, max_height=12, figpadding=0.2,
-    attribute_to_plot='net_flux', show_frame=False, xticks=False, yticks=False, ax=None,
+    show_frame=False, xticks=False, yticks=False, ax=None,
     **textkwargs):
     r"""Network representation of given matrix